Moved main sweeper code to _impl.py so that it doesn't get accidentally loaded along with all its monkeypatching when the user doesn't want to use the sweeper. Upon running a task function, hydra loads all available plugins, even those that won't be used. Improved code typing. Added implicit start to yamlfmt pre-commit since it clashed with hydra's # @package _global_ indicator.
