+++ example/conf/sweeper/wandb.yaml
@@ -1,4 +1,3 @@
----
 # @package _global_
 defaults:
   - override /hydra/sweeper: wandb
